requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pynacl==1.5.0
pandas==2.1.1
matplotlib==3.8.0
